+++ .dotnet/CHANGELOG.md
@@ -23,6 +23,8 @@
 - Changed `GeneratedImageQuality` from an enum to an "extensible enum". (commit_hash)
 - Changed `GeneratedImageStyle` from an enum to an "extensible enum". (commit_hash)
 - Removed method overloads in `AssistantClient` and `VectorStoreClient` that take complex parameters in favor of methods that take simple string IDs. (commit_hash)
+- Updated the `TokenIds` property type in the `TranscribedSegment` type from `IReadOnlyList<int>` to `ReadOnlyMemory<int>`. (commit_hash)
+- Updated the `inputs` parameter type in the `GenerateEmbeddings` and `GenerateEmbeddingsAsync` methods of `EmbeddingClient` from `IEnumerable<IEnumerable<int>>` to `IEnumerable<ReadOnlyMemory<int>>`. (commit_hash)
 
 ### Bugs Fixed